Cambio de color de los HTML y se implementa el modulo de medico
